vel_to_hover node

+++ setup.sh
@@ -1,17 +1,22 @@
 #!/usr/bin/env bash
 set -e
 
+# Check if running inside Docker
+if [ ! -f "/.dockerenv" ]; then
+  echo "❌ This script must be run inside the Docker container."
+  exit 1
+fi
+
 echo "==> Setting up Crazyswarm2"
 
-ROOT_DIR=/root
-ROS2_WS=${ROOT_DIR}/ros2_ws
+ROS2_WS=${HOME}/ros2_ws
 
 # Import repositories
 echo "==> Cloning repo"
 
 mkdir -p ${ROS2_WS}/src
 
-if [ ! -d "${ROS2_WS}/src/crazyswarm2" ]; then
+if [ ! -d "${ROS2_WS}/src/crazyswarm2/crazyflie" ]; then
   git -C ${ROS2_WS}/src \
     clone --recursive \
     https://github.com/IMRCLab/crazyswarm2
@@ -27,3 +32,4 @@ colcon build \
   --cmake-args -DCMAKE_BUILD_TYPE=Release
 
 echo "==> Setup completed successfully ✅"
+
